requirements-shib.txt
Flask==3.0.0
Flask-Cors==4.0.0
pysaml2==7.5.0
xmlsec==1.3.12
lxml==5.1.0
pyjwt==2.8.0
requests==2.31.0
gunicorn==21.2.0
